Move functions to show chart into chart.py
